Move expiration date coloring and icon code into util.php
These functions were getting redclared a in tool-slices and in
slice.php, keep them all in util.php now.
